Tests/LibWeb/Text/expected/HTML/WindowOrWorkerGlobalScope-reportError.txt

message = Reporting an Error!
filename = 
lineno = 0
colno = 0
error = Error: Reporting an Error!
